Dead Code Cleanup & Controller Pattern Refactoring
