+++ backends/advanced/scripts/enroll_speaker_service.py
