+++ backends/advanced/src/advanced_omi_backend/controllers/__init__.py
@@ -1,0 +1,13 @@
+"""
+Controllers for handling business logic separate from route definitions.
+"""
+
+from . import memory_controller, user_controller, conversation_controller, client_controller, system_controller
+
+__all__ = [
+    "memory_controller",
+    "user_controller", 
+    "conversation_controller",
+    "client_controller",
+    "system_controller"
+]
